+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- added cgroup_ns option to container resource with default private
+
 ## 11.1.1 - *2023-07-10*
 
 ## 11.1.0 - *2023-06-06*