+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- fix the generation of the docker version string to install on ubuntu jammy
+
 ## 11.2.0 - *2023-08-10*
 
 - added cgroup_ns option to container resource with default private